refactor: remove theme prop drilling
